rename plot submodule
