+++ hazGAN/plots.py
